# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cybersecurity Platform API
  version: 1.0.0
  description: API contract for public intake, secure client portal, and role-protected administration.
servers:
  - url: /api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: portal_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
        requestId: { type: string }
    Reference:
      type: object
      properties:
        ok: { const: true }
        data:
          type: object
          required: [reference, status]
          properties:
            reference: { type: string }
            status: { type: string }
paths:
...
  /forms/contact:
    post:
      summary: Submit an idempotent contact request
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, workEmail, organisation, topic, message, consent, idempotencyKey]
              properties:
                name: { type: string, maxLength: 120 }
                workEmail: { type: string, format: email }
                organisation: { type: string, maxLength: 160 }
                topic: { enum: [general, services, partnership, privacy, other] }
                message: { type: string, maxLength: 4000 }
                consent: { const: true }
                idempotencyKey: { type: string, format: uuid }
      responses:
        '201':
          {
            description: Persisted,
            content: { application/json: { schema: { $ref: '#/components/schemas/Reference' } } },
          }
        '422':
          {
            description: Validation error,
            content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } },
          }
  /forms/assessment:
    post:
      summary: Submit an assessment scope request
      responses: { '201': { description: Persisted }, '422': { description: Validation error } }
  /forms/incident:
    post:
      summary: Submit an incident response request
      responses: { '201': { description: Persisted }, '422': { description: Validation error } }
  /forms/vulnerability:
    post:
      summary: Submit a responsible disclosure report
      responses: { '201': { description: Persisted }, '422': { description: Validation error } }
...
  /auth/login:
    post:
      summary: Create access and rotating refresh sessions
      responses:
        {
          '200': { description: Authenticated },
          '401': { description: Generic invalid credentials response },
        }
  /auth/refresh:
    post:
      summary: Rotate a valid refresh session
      responses: { '200': { description: Rotated }, '401': { description: Invalid or revoked } }
  /auth/logout:
    post:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Revoke current session
      responses: { '200': { description: Revoked } }
...